feat: sqlite output, #MELA2-26 (#4)
Simulation state tree is now written to an SQLite database. The legacy
state output mechanism is still temporarily available but is going to be
removed in the near future. All post-processing and export operations
have also been removed. Any manipulation or reporting of simulation
results is to be carried out with external tools using the simulation
output from the database.

ver: update version and changelog